# docs/swagger.yaml
basePath: /
definitions:
...
  response.Response:
    properties:
      code:
        description: 提示代码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
host: localhost:8082
info:
  contact:
    name: API Support
    url: https://github.com/carrot-hu23/dst-admin-go
  description: 饥荒联机版服务器管理后台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DST Admin Go API
  version: "1.7.0"
paths:
...
  /api/game/level/server/download:
    get:
      description: 下载指定世界的完整服务器日志文件
      parameters:
      - description: 集群名称
        in: query
        name: clusterName
        type: string
      - description: 世界名称
        in: query
        name: levelName
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 服务器日志文件
          schema:
            type: file
      summary: 下载服务器日志
      tags:
      - log
  /api/game/level/server/log:
    get:
      description: 获取指定世界的服务器日志（默认最近100行）
      parameters:
      - description: 集群名称
        in: query
        name: clusterName
        type: string
      - description: 世界名称
        in: query
        name: levelName
        required: true
        type: string
      - description: 返回日志行数，默认为100
        in: query
        name: lines
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 获取服务器日志
      tags:
      - log
  /api/game/log/stream:
    get:
      consumes:
      - text/event-stream
      description: 获取指定世界的实时日志流 (SSE)
      parameters:
      - description: 集群名称
        in: query
        name: clusterName
        type: string
      - description: 世界名称
        in: query
        name: levelName
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE 格式的日志流
          schema:
            type: string
      summary: 服务器日志流
      tags:
      - log
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"